+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/Helpers/EmbeddingContentPartGraphSyncer.cs
@@ -2,9 +2,10 @@
 using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces.Contexts;
 using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces.EmbeddedContentItemsGraphSyncer;
 using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces.Results.AllowSync;
+using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Parts;
 using Newtonsoft.Json.Linq;
 
-namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Parts
+namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Helpers
 {
     public abstract class EmbeddingContentPartGraphSyncer : ContentPartGraphSyncer
     {
